+++ back_end/saolei/videomanager/config.py
@@ -1,0 +1,1 @@
+PLUCK_TIMEOUT_SECONDS = 60
