+++ back_end/saolei/videomanager/tasks.py
@@ -1,21 +1,7 @@
-import math
-
 from django.tasks import task
-from timeout_decorator import timeout
 
-from utils.parser import create_video_from_data
 from .models import VideoModel
-
-PLUCK_TIMEOUT_SECONDS = 60
-
-
-def normalize_pluck(value) -> float | None:
-    if value is None:
-        return None
-    value = float(value)
-    if not math.isfinite(value) or value <= 0:
-        return None
-    return value
+from .utils import calculate_pluck, normalize_pluck
 
 
 def helper_video_pluck(video: VideoModel):
@@ -23,19 +9,6 @@ def helper_video_pluck(video: VideoModel):
 
     if is_custom_pluck_video(video) and video.pluck is None:
         task_video_pluck.enqueue(video.id)
-
-
-@timeout(PLUCK_TIMEOUT_SECONDS, use_signals=False, timeout_exception=TimeoutError)
-def calculate_pluck(file_path: str):
-    with open(file_path, 'rb') as file:
-        data = file.read()
-
-    parsed_video, _ = create_video_from_data(file_path, data)
-    parsed_video.parse()
-    parsed_video.analyse()
-    parsed_video.analyse_for_features(['pluck'])
-    parsed_video.current_time = 1e8
-    return parsed_video.pluck
 
 
 @task(priority=-1)
